feat/ remove all whitelist code
