+++ bat-utils/lib/hapi-rate-limiter.js
@@ -6,24 +6,8 @@ const {
 const _ = require('underscore')
 const underscore = _
 
-const whitelist = require('./hapi-auth-whitelist')
-
 const pluginName = 'rateLimitRedisPlugin'
-
-const graylist = {
-  addresses: process.env.IP_GRAYLIST && process.env.IP_GRAYLIST.split(',')
-}
-
-if (graylist.addresses) {
-  graylist.authorizedAddrs = []
-  graylist.authorizedBlocks = []
-
-  graylist.addresses.forEach((entry) => {
-    if ((entry.indexOf('/') === -1) && (entry.split('.').length === 4)) return graylist.authorizedAddrs.push(entry)
-
-    graylist.authorizedBlocks.push(new Netmask(entry))
-  })
-}
+const braveHapi = require('./extras-hapi');
 
 module.exports = (runtime) => {
   const redisClient = (runtime.cache && runtime.cache.cache)
@@ -38,13 +22,6 @@ module.exports = (runtime) => {
     redis: redisClient,
     keyPrefix: 'rate-limiter-authed',
     points: 3000, // requests per
-    duration: 60 // seconds by IP
-  })
-
-  const rateLimiterWhitelisted = new RateLimiterRedis({
-    redis: redisClient,
-    keyPrefix: 'rate-limiter-whitelist',
-    points: 60000, // requests per
     duration: 60 // seconds by IP
   })
 
@@ -74,7 +51,6 @@ module.exports = (runtime) => {
     redisClient,
     rateLimiter,
     rateLimiterAuthed,
-    rateLimiterWhitelisted,
     noRateLimiter
   }
 
@@ -114,24 +90,8 @@ module.exports = (runtime) => {
       if (process.env.NODE_ENV !== 'production') {
         return internals.noRateLimiter
       }
-      const ipaddr = whitelist.ipaddr(request)
-      if (ipaddr === '127.0.0.1') {
-        return internals.noRateLimiter
-      }
 
-      if ((graylist.authorizedAddrs) &&
-          ((graylist.authorizedAddrs.indexOf(ipaddr) !== -1) ||
-           (underscore.find(graylist.authorizedBlocks, (block) => { return block.contains(ipaddr) })))) {
-        return internals.noRateLimiter
-      }
-
-      if (whitelist.authorizedP(ipaddr)) {
-        if (request.auth && request.auth.credentials && request.auth.credentials.token && request.auth.credentials.scope) {
-          return internals.rateLimiterWhitelisted
-        }
-
-        return internals.rateLimiterAuthed
-      }
+      return internals.rateLimiterAuthed
     } catch (e) {}
 
     return internals.rateLimiter
@@ -139,7 +99,7 @@ module.exports = (runtime) => {
 
   function rateLimitKey (request) {
     try {
-      return whitelist.ipaddr(request) + ':' + runtime.config.server.host
+      return braveHapi.ipaddr(request) + ':' + runtime.config.server.host
     } catch (e) {
       return 'default'
     }
